+++ integration_test/test_IOMULT.go
@@ -2,8 +2,6 @@ package main
 
 import (
 	"fmt"
-	"log"
-	"os"
 )
 
 // 複数クライアント(A, B, C)にコネクションと3分割したメッセージを用意して, ランダムに送信する
@@ -65,26 +63,4 @@ func testIOMULT() {
 		}
 		return true, nil
 	})
-}
-
-// for color print
-const (
-	red   = "\033[31m"
-	green = "\033[32m"
-	reset = "\033[0m"
-)
-
-// テスト関数を渡してその結果に合わせたメッセージを出力する関数です
-func testHandler(name string, test func() (bool, error)) {
-	fmt.Print("[ " + name + " ] ")
-	ok, err := test()
-	if err != nil {
-		log.Fatalf("erro occured!: %v", err)
-	}
-	if ok {
-		fmt.Println(green, "ok", reset)
-	} else {
-		fmt.Println(red, "error", reset)
-		os.Exit(1) // TODO: テスト全体でエラーがあれば最後にexit(1)する
-	}
 }
